[pax map] pax map cycle delta misc fixes

+++ airline-data/src/main/scala/com/patson/data/ConsumptionHistorySource.scala
@@ -1,11 +1,8 @@
 package com.patson.data
 import com.patson.data.Constants._
+import com.patson.model._
 
 import scala.collection.mutable.ListBuffer
-import com.patson.model._
-import java.util
-
-import com.mysql.jdbc.exceptions.MySQLDataException
 
 
 object ConsumptionHistorySource {
@@ -59,11 +56,9 @@ object ConsumptionHistorySource {
           }
         val toTableName = PASSENGER_HISTORY_TABLE + "_" + i
 
-        connection.createStatement().executeUpdate(s"DROP TABLE IF EXISTS $toTableName")
-        try {
+        if (Meta.isTableExist(connection, fromTableName)) {
+          connection.createStatement().executeUpdate(s"DROP TABLE IF EXISTS $toTableName")
           connection.createStatement().executeUpdate(s"ALTER TABLE $fromTableName RENAME $toTableName")
-        } catch  {
-          case e : java.sql.SQLException => println(s"Skipping $fromTableName/$toTableName message: ${e.getMessage}")
         }
       }
 
@@ -246,64 +241,68 @@ object ConsumptionHistorySource {
       case Some(link) =>
         val connection = Meta.getConnection()
         try {
-          val preparedStatement = connection.prepareStatement("SELECT route_id FROM " + tableName + " WHERE link = ? ")
+          if (Meta.isTableExist(connection, tableName)) {
+            val preparedStatement = connection.prepareStatement("SELECT route_id FROM " + tableName + " WHERE link = ? ")
 
-          preparedStatement.setInt(1, linkId)
-          val resultSet = preparedStatement.executeQuery()
+            preparedStatement.setInt(1, linkId)
+            val resultSet = preparedStatement.executeQuery()
 
-          val relatedRouteIds = new ListBuffer[Int]()
-          while (resultSet.next()) {
-            relatedRouteIds += resultSet.getInt("route_id")
-          }
+            val relatedRouteIds = new ListBuffer[Int]()
+            while (resultSet.next()) {
+              relatedRouteIds += resultSet.getInt("route_id")
+            }
 
-          if (relatedRouteIds.isEmpty) {
-            Map.empty
+            if (relatedRouteIds.isEmpty) {
+              Map.empty
+            } else {
+              val queryString = new StringBuilder("SELECT * FROM " + tableName + " where route_id IN (");
+              for (i <- 0 until relatedRouteIds.size - 1) {
+                queryString.append("?,")
+              }
+              queryString.append("?)")
+
+              val relatedRouteStatement = connection.prepareStatement(queryString.toString())
+
+              for (i <- 0 until relatedRouteIds.size) {
+                relatedRouteStatement.setInt(i + 1, relatedRouteIds(i))
+              }
+
+              val relatedRouteSet = relatedRouteStatement.executeQuery()
+
+              val linkConsiderationsByRouteId = scala.collection.mutable.Map[Int, ListBuffer[LinkConsideration]]()
+              val routeConsumptions = new scala.collection.mutable.HashMap[Int, (PassengerType.Value, Int)]()
+
+              val relatedLinkIds = scala.collection.mutable.HashSet[Int]()
+              while (relatedRouteSet.next()) {
+                relatedLinkIds += relatedRouteSet.getInt("link")
+              }
+              val linkMap = LinkSource.loadLinksByIds(relatedLinkIds.toList).map(link => (link.id, link)).toMap
+
+              relatedRouteSet.beforeFirst()
+              while (relatedRouteSet.next()) {
+                val routeId = relatedRouteSet.getInt("route_id")
+                val passengerType = PassengerType.apply(relatedRouteSet.getInt("passenger_type"))
+                val passengerCount = relatedRouteSet.getInt("passenger_count")
+                val relatedLinkId = relatedRouteSet.getInt("link")
+                val relatedLink = linkMap.getOrElse(relatedLinkId, Link.fromId(relatedLinkId))
+                val linkConsideration = new LinkConsideration(relatedLink, 0, LinkClass.fromCode(relatedRouteSet.getString("link_class")), relatedRouteSet.getBoolean("inverted"))
+
+                val existingConsiderationsForThisRoute = linkConsiderationsByRouteId.getOrElseUpdate(routeId, ListBuffer[LinkConsideration]())
+
+                existingConsiderationsForThisRoute += linkConsideration
+                routeConsumptions.put(routeId, (passengerType, passengerCount))
+              }
+
+              val result = linkConsiderationsByRouteId.map {
+                case (routeId: Int, considerations: ListBuffer[LinkConsideration]) => (new Route(considerations.toList, 0, routeId), routeConsumptions(routeId))
+              }.toMap
+
+              println("Loaded " + result.size + " routes related to link " + link)
+
+              result
+            }
           } else {
-            val queryString = new StringBuilder("SELECT * FROM " + PASSENGER_HISTORY_TABLE + " where route_id IN (");
-            for (i <- 0 until relatedRouteIds.size - 1) {
-              queryString.append("?,")
-            }
-            queryString.append("?)")
-
-            val relatedRouteStatement = connection.prepareStatement(queryString.toString())
-
-            for (i <- 0 until relatedRouteIds.size) {
-              relatedRouteStatement.setInt(i + 1, relatedRouteIds(i))
-            }
-
-            val relatedRouteSet = relatedRouteStatement.executeQuery()
-
-            val linkConsiderationsByRouteId = scala.collection.mutable.Map[Int, ListBuffer[LinkConsideration]]()
-            val routeConsumptions = new scala.collection.mutable.HashMap[Int, (PassengerType.Value, Int)]()
-
-            val relatedLinkIds = scala.collection.mutable.HashSet[Int]()
-            while (relatedRouteSet.next()) {
-              relatedLinkIds += relatedRouteSet.getInt("link")
-            }
-            val linkMap = LinkSource.loadLinksByIds(relatedLinkIds.toList).map(link => (link.id, link)).toMap
-
-            relatedRouteSet.beforeFirst()
-            while (relatedRouteSet.next()) {
-              val routeId = relatedRouteSet.getInt("route_id")
-              val passengerType = PassengerType.apply(relatedRouteSet.getInt("passenger_type"))
-              val passengerCount = relatedRouteSet.getInt("passenger_count")
-              val relatedLinkId = relatedRouteSet.getInt("link")
-              val relatedLink = linkMap.getOrElse(relatedLinkId, Link.fromId(relatedLinkId))
-              val linkConsideration = new LinkConsideration(relatedLink, 0, LinkClass.fromCode(relatedRouteSet.getString("link_class")), relatedRouteSet.getBoolean("inverted"))
-
-              val existingConsiderationsForThisRoute = linkConsiderationsByRouteId.getOrElseUpdate(routeId, ListBuffer[LinkConsideration]())
-
-              existingConsiderationsForThisRoute += linkConsideration
-              routeConsumptions.put(routeId, (passengerType, passengerCount))
-            }
-
-            val result = linkConsiderationsByRouteId.map {
-              case (routeId: Int, considerations: ListBuffer[LinkConsideration]) => (new Route(considerations.toList, 0, routeId), routeConsumptions(routeId))
-            }.toMap
-
-            println("Loaded " + result.size + " routes related to link " + link);
-
-            result
+            Map.empty
           }
         } finally {
           connection.close()
